runtests.sh: set up a new postgresql cluster and run tests there
This allows running tests without installing plproxy and should not conflict
with any other PostgreSQL instances.  There are a couple of hacks in
runtests.sh to work around limitations in installing extensions from outside
the global postgresql share directory.

Also reformatted all connection strings in existing test cases to set the
target hostname as the first field so we can rewrite it when running tests.

Signed-off-by: Oskari Saarenmaa <os@ohmu.fi>